Backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-allauth
djangorestframework-simplejwt
python-dotenv
openai
razorpay
pydub
mutagen
hypothesis
pytest
pytest-django
factory-boy
responses

# Production improvements
celery==5.3.6
redis==5.0.1
django-ratelimit==4.1.0
